requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic-settings>=2.10.1
starlett=0.47.1
httpx==0.28.1
